add orders to mysql & fix create order bug

+++ cfm-node/src/models/order.model.js
@@ -16,7 +16,7 @@ module.exports = (sequelize, Sequelize) => {
         type: Sequelize.STRING
       },
       totalprice:{
-        type: Sequelize.DOUBLE
+        type: Sequelize.FLOAT
       }
     });
     return Orders;
